requirements/base.txt
django
django-cors-headers
django-constance
django-filter
djangorestframework
drf-nested-routers
gunicorn
python-dotenv
